+++ capreolus/collection/msmarco.py
@@ -1,0 +1,10 @@
+from .base import Collection
+from capreolus import ConfigOption
+
+# @Collection.register
+# class MSMarco(Collection):
+#     module_name = "msmarco"
+#     config_keys_not_in_path = ["path"]
+#     collection_type = "TrecCollection"
+#     generator_type = "DefaultLuceneDocumentGenerator"
+#     config_spec = [ConfigOption("path", "/GW/NeuralIR/nobackup/msmarco/trec_format", "path to corpus")]
